Fix infinite loop
If the number of students is great than the number of saints, the app enters an infinite loop. My temporary fix is to change the number entered to array.length whenever the number of students is great than the number of saints. It's far from ideal, but at least the app doesn't break.

I think what would be ideal is if the loop simply started over and allowed the reusing of duplicates. That way, the students can double up if necessary.