Update all lists item delete behavior
Before, it rendered all lists. This was to make sure the options updated as well. But then the details element closed, because all lists was also being rerendered. Duh! So now it just calls the render options function. No downsides that I can see, though that doesn't mean much, does it?