Render pairings as a single list
Before, we were rendering two separate lists side by side. But that wasn't working so well on mobile because text-wrapping threw things out of alignment. So now the pairings are rendered in a single list.

One potential issue is what to do with extras. So if one list is longer than another, what do we do with the list items that are left over? Right now, we aren't rendering them at all.