Add save button to saved lists
Now when you make changes (add or delete items) you hit save the list rerenders. We're doing it this way because beforehand the list didn't sort automatically and the delete buttons weren't rendered when you added items. That's a problem because on mobile, the lists wouldn't sort until the user quit the app, and I don't want to rely on that.