Remove PWA stuff
Apparently this is experimental tech, which maybe is why it was proving hard to work with. Either way, it works in Chromium. I just don't have any control over how those native features are triggered.

If I can figure it out, what I'll probably do is put it on the app store, then link to it.